projects: Force local state storage for now
For the sake of prototyping real operations, for now we'll just force all
workspaces to use local state files in a prototype-specific directory.
Eventually we should make this customizable using either state_storage
or remote configuration, but we don't have the necessary remote client
APIs for that yet (existing packages are built around the old backend
model) and this temporary solution allows for some forward progress
without significant refactoring of the state storage mechanisms.